refactor: unify card sizes, improve chat UX, restructure ChatMessage component
